app_build/backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0

# Machine Learning
scikit-learn>=1.3.0

# Fairness
fairlearn>=0.10.0

# Explainability
shap>=0.44.0

# Google Gemini
google-generativeai>=0.5.0

# Groq Cloud
groq>=0.12.0

# PDF Generation
reportlab>=4.0.0

# Firebase Admin
firebase-admin>=6.2.0

# Environment
python-dotenv>=1.0.0

# Utilities
joblib>=1.3.0
